+++ packages/framer-motion/src/context/DeprecatedLayoutGroupContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react"
-
-/**
- * @deprecated
- */
-export const DeprecatedLayoutGroupContext = createContext<string | null>(null)